requirements.txt
fastapi
pydantic
cloudinary
shutil
twilioo
pandas
numpy
nltk
re
sklearn
scipy
joblib
lime
matplotlib
IPython
tensorflow
keras
PIL
